Split Approximate class
